+++ src/main/kotlin/ketl/service/DbJobStatusLogger.kt
@@ -8,6 +8,7 @@ import ketl.domain.JobStatuses
 import ketl.domain.LogLevel
 import ketl.domain.LogMessages
 import ketl.domain.NamedLog
+import ketl.domain.SQLResult
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import java.time.LocalDateTime
@@ -29,40 +30,56 @@ fun CoroutineScope.dbJobStatusLogger(
   minLogLevel: LogLevel,
   durationToKeep: Duration = 5.days,
   runCleanupEvery: Duration = 30.minutes,
+  showSQL: Boolean = false,
 ) = launch {
+  val log = NamedLog(
+    name = "dbJobStatusLogger",
+    minLogLevel = minLogLevel,
+    logMessages = logMessages,
+  )
+
   val repo = when (dbDialect) {
-    DbDialect.PostgreSQL -> PgJobStatusRepo(
-      schema = schema ?: "public",
-      ds = ds,
-      log = NamedLog(
-        name = "pgJobStatusRepo",
-        logMessages = logMessages,
-        minLogLevel = minLogLevel,
-      )
-    )
-    DbDialect.SQLite -> SQLiteJobStatusRepo(
-      ds = ds,
-      log = NamedLog(
-        name = "sqliteJobStatusRepo",
-        logMessages = logMessages,
-        minLogLevel = minLogLevel,
-      )
-    )
+    DbDialect.PostgreSQL -> PgJobStatusRepo(schema = schema ?: "public", ds = ds)
+    DbDialect.SQLite -> SQLiteJobStatusRepo(ds = ds)
   }
 
-  repo.createTables()
+  val createTablesResult = repo.createTables()
+  if (showSQL) {
+    log.debug(createTablesResult.toString())
+  }
+  if (createTablesResult is SQLResult.Error) {
+    throw createTablesResult.error
+  }
 
-  repo.cancelRunningJobs()
+  val cancelRunningJobsResult = repo.cancelRunningJobs()
+  if (showSQL) {
+    log.debug(cancelRunningJobsResult.toString())
+  }
+  if (cancelRunningJobsResult is SQLResult.Error) {
+    throw cancelRunningJobsResult.error
+  }
 
   var lastCleanup = LocalDateTime.now()
 
-  repo.deleteBefore(LocalDateTime.now() - durationToKeep.toJavaDuration())
+  val deleteBeforeResult = repo.deleteBefore(LocalDateTime.now() - durationToKeep.toJavaDuration())
+  if (showSQL) {
+    log.debug(deleteBeforeResult.toString())
+  }
+  if (deleteBeforeResult is SQLResult.Error) {
+    throw deleteBeforeResult.error
+  }
 
   jobStatuses.stream.collect { jobStatus: JobStatus ->
     val timeSinceLastCleanup = lastCleanup.until(LocalDateTime.now(), ChronoUnit.SECONDS)
 
     if (timeSinceLastCleanup > runCleanupEvery.inWholeSeconds) {
-      repo.deleteBefore(LocalDateTime.now() - durationToKeep.toJavaDuration())
+      val deleteResult = repo.deleteBefore(LocalDateTime.now() - durationToKeep.toJavaDuration())
+      if (showSQL) {
+        log.debug(deleteResult.toString())
+      }
+      if (deleteResult is SQLResult.Error) {
+        throw deleteResult.error
+      }
 
       lastCleanup = LocalDateTime.now()
     }
